start naive planning based on sampled region
started as a test for moving the code into a real moddule and different
smaller files. This seems to work now
Implement naive planning, where we sample a region around the viewpoint
that spans the admissible movement range, compute the potential for each
vp candidate and pick the argmax.
The parameters need to be tuned, the fov is very large and produces
weird results that look to good.